Move redstone logic related functions to own util file
